shell-ini-config: fix subtle section parsing bug
The shell-ini-config function was broken in a subtle way. Once we
encounter [section] tag that matched the requested section, we enabled
reading each line and returning the config setting that matched.
In the case where the requested section did not have a defined value,
but the value was defined in a section below the requested section in
the configuration file, we erroneously return the value from a different
section.

This patch fixes the bug by disabling section processing whenever we
encounter a new [section] tag. This fix also allows the same section
multiple times, which is probably not very useful but shouldn't cause
any issues.

Signed-off-by: Jacob Keller <jacob.e.keller@intel.com>