ip-address: Add ipv6_addr_type
This function interprets the given option value as an IPv6 address
similarly to inet_pton(3), and determines its type in regard to various
special properties. If the input string is a well-formed IPv6 address,
the function returns successfully.

The intended purpose of `ipv6_addr_type` is to prevent misuse of
special-use addresses. For example, since link-local unicast addresses
only make sense within a link and are incomplete without a scope
identifier, it generally makes no sense to specify them in DNS AAAA
records. Similar reasoning can be applied to multicast addresses, etc.

Signed-off-by: Arseny Maslennikov <ar@cs.msu.ru>