ip-address: Add ipv6_ptonx
This function takes a string representation of an IPv6 address and
prints out its raw, network byte order representation as a sequence
of 16 adjacent 2-digit hexadecimal octets.

The input string is parsed in accordance with the inet_pton(3)
algorithm. This means that we do not support scope identifiers (a
percent sign, followed by a network interface ID), nor common transport
layer service port designations (like a colon followed by an unsigned
16-bit decimal port number), as well as any trailing characters after
the IPv6 address. The input string is not meant to be a sockaddr, just
a network address.

While this provides an easy way to perform arbitrary bitwise operations
on the input address, a true `*_pton` function which writes out raw
bytes would have little use in a shell environment.

Signed-off-by: Arseny Maslennikov <ar@cs.msu.ru>