ip-address: Add ipv4_ptonx
This function takes a string representation of an IPv4 address and
prints out its raw, network byte order representation as a sequence
of 4 adjacent 2-digit hexadecimal octets.

The input string is parsed in accordance with the inet_pton(3)
algorithm. This means that we do not support some peculiar notations
recognized by inet_aton(3) and, by extension, getaddrinfo(3). An input
string with trailing characters after the address is also considered
malformed.

While this provides an easy way to perform arbitrary bitwise operations
on the input address, a true `*_pton` function which writes out raw
bytes would have little use in a shell environment.

Signed-off-by: Arseny Maslennikov <ar@cs.msu.ru>