ip-address: Add tests for ipv4_prefix2mask
Add some tests that cover all the possible octet values in the mask,
as well as invalid prefix lengths.

While running the tests, it was discovered the function did not work
on mksh at all:
- that shell stores integers as int32_t in arith context, so numbers
  like 0x80000000 are < 0, and bit shift to the right has been observed
  to sign-extend;
- that shell cannot left-shift an integer by 32 bits (and evaluate to a
  multiple of 4294967296), since the shift amount is considered modulo 32.
We have to replace the algorithm to not rely on left shift by 32, or any
right shift. We move it to a separate function to be used from more
places in subsequent changes.

Signed-off-by: Arseny Maslennikov <ar@cs.msu.ru>